+++ test/events-test.js
@@ -52,4 +52,47 @@ describe('Tokenizer Events', function () {
       })
     })
   })
+
+  describe('[seek]', function () {
+    var p = new Tokenizer(options)
+    var flag = false
+
+    it('should emit [seek] on `seek()`', function (done) {
+      p.addRule(' ', function (token, idx, type) {
+        assert.deepEqual(flag, true)
+        done()
+      })
+      p.addRule(1, function () {
+        p.seek(1)
+      })
+      p.on('seek', function (i) {
+        assert.equal(i, 1)
+        flag = true
+      })
+      p.write('abc  ')
+    })
+  })
+
+  describe('[loadruleset]', function () {
+    var p = new Tokenizer(options)
+    var flag = false
+
+    it('should emit [loadruleset] when a rule is loaded', function (done) {
+      p.addRule(1, function () {
+        p.loadRuleSet('ruleSet2')
+      })
+      p.saveRuleSet('ruleSet1')
+      p.addRule(0, function (token, idx, type) {
+        assert.deepEqual(flag, true)
+        done()
+      })
+      p.saveRuleSet('ruleSet2')
+      p.loadRuleSet('ruleSet1')
+      p.on('loadruleset', function (ruleset) {
+        assert.equal(ruleset, 'ruleSet2')
+        flag = true
+      })
+      p.write('abc')
+    })
+  })
 })