+++ test/rules-properties-methods-test.js
@@ -8,30 +8,61 @@ var options = {}
 
 describe('Tokenizer Properties Methods', function () {
   describe('#ignore', function () {
-    var p = new Tokenizer(options)
-    it('should ignore the token matching the rule', function (done) {
-      p.ignore(true)
-      p.addRule('a', function (token, idx, type) {
-        done( new Error('should not trigger') )
+    describe('on non empty buffer', function () {
+      var p = new Tokenizer(options)
+      it('should ignore the token matching the rule', function (done) {
+        p.ignore(true)
+        p.addRule('a', function (token, idx, type) {
+          done( new Error('should not trigger') )
+        })
+        p.ignore()
+        p.addRule('b', function (token, idx, type) {
+          done()
+        })
+        p.write('ab')
       })
-      p.ignore()
-      p.addRule('b', function (token, idx, type) {
+    })
+
+    // This makes non sense but is possible...
+    describe('on empty buffer', function () {
+      var p = new Tokenizer(options)
+      var flag = false
+
+      it('should ignore the handler', function (done) {
+        p.ignore(true)
+        p.addRule(1, 'consume data')
+        p.addRule(0, function (token, idx, type) {
+          flag = true
+        })
+        p.write('a')
+        assert.deepEqual(flag, false)
         done()
       })
-      p.write('ab')
     })
   })
 
   describe('#quiet', function () {
-    describe('on Number rule', function () {
+    describe('on last Number rule', function () {
       var p = new Tokenizer(options)
-      it('should also apply', function (done) {
+      it('should give the token size', function (done) {
         p.quiet(true)
         p.addRule(1, function (token, idx, type) {
           assert.equal(token, 1)
           done()
         })
         p.write('a')
+      })
+    })
+
+    describe('on non last Number rule', function () {
+      var p = new Tokenizer(options)
+      it('should not apply', function (done) {
+        p.quiet(true)
+        p.addRule(3, 'c', function (token, idx, type) {
+          assert.equal(token, 'abc')
+          done()
+        })
+        p.write('abc')
       })
     })
 
@@ -143,6 +174,97 @@ describe('Tokenizer Properties Methods', function () {
         })
         p.write('aa')
       })
+    })
+
+    describe('while paused', function () {
+      var p = new Tokenizer(options)
+      it('should continue at the indexed rule', function (done) {
+        p.continue(1)
+        p.addRule('a', function (token, idx, type) {
+          p.pause()
+        })
+        p.continue()
+        p.addRule(' ', 'no match')
+        p.addRule('b', function (token, idx, type) {
+          done()
+        })
+        p.write('ab')
+        p.resume()
+      })
+    })
+  })
+
+  describe('#saveProps', function () {
+    var p = new Tokenizer(options)
+    var props = []
+
+    for (prop in p)
+      if (p.hasOwnProperty(prop) && /^_p_/.test(prop)) props.push( prop.substr(3) )
+
+    it('should save the current properties', function (done) {
+      var saved = {}
+      
+      // Change all properties
+      for (var prop, i = 0, n = props.length; i < n; i++) {
+        prop = props[i]
+        switch (prop) {
+          case 'escape':
+            p.escaped('\\')
+            saved.escape = '\\'
+          break
+          case 'continue':
+            p.continue(123)
+            saved.continue = 123
+          break
+          case 'next':
+            p.next('ruleSet')
+            saved.next = 'ruleSet'
+          break
+          default:
+            p[ prop ](true)
+            saved[prop] = true
+        }
+      }
+      p.saveProps()
+
+      assert.deepEqual(p.savedProps, saved)
+      done()
+    })
+  })
+
+  describe('#loadProps', function () {
+    var p = new Tokenizer(options)
+    var props = []
+
+    for (prop in p)
+      if (p.hasOwnProperty(prop) && /^_p_/.test(prop)) props.push( prop.substr(3) )
+
+    it('should load the current properties', function (done) {
+      p.saveProps()
+
+      var saved = p.savedProps
+      
+      // Change all properties
+      for (var prop, i = 0, n = props.length; i < n; i++) {
+        prop = props[i]
+        switch (prop) {
+          case 'escape':
+            p.escaped('\\')
+          break
+          case 'continue':
+            p.continue(123)
+          break
+          case 'next':
+            p.next('ruleSet')
+          break
+          default:
+            p[ prop ](true)
+        }
+      }
+      p.loadProps()
+
+      assert.deepEqual(p.savedProps, saved)
+      done()
     })
   })
 })